+++ runtime/vm/service.h
@@ -19,7 +19,7 @@
 namespace dart {
 
 #define SERVICE_PROTOCOL_MAJOR_VERSION 4
-#define SERVICE_PROTOCOL_MINOR_VERSION 17
+#define SERVICE_PROTOCOL_MINOR_VERSION 18
 
 class Array;
 class EmbedderServiceHandler;