+++ pkg/analyzer/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 8.0.0-dev
 * Remove deprecated element model V1.
+* Remove deprecated `DartType.isStructurallyEqualTo`.
 
 ## 7.4.1
 * Restore `InstanceElement.augmented` getter.